+++ src/modules/commerce/module.ts
@@ -1,6 +1,6 @@
 import { FakerError } from '../../errors/faker-error';
 import { ModuleBase } from '../../internal/module-base';
-import { calculateUPCCheckDigit } from './upc-check-digit';
+import { calculateUPCCheckDigit } from './_upc-check-digit';
 
 // Source for official prefixes: https://www.isbn-international.org/range_file_generation
 const ISBN_LENGTH_RULES: Record<
