+++ src/modules/helpers/module.ts
@@ -3,8 +3,8 @@ import type { Faker } from '../../faker';
 import { SimpleModuleBase } from '../../internal/module-base';
 import type { SimpleFaker } from '../../simple-faker';
 import type { NumberRange } from '../../utils/types';
-import { fakeEval } from './eval';
-import { luhnCheckValue } from './luhn-check';
+import { fakeEval } from './_eval';
+import { luhnCheckValue } from './_luhn-check';
 
 /**
  * Returns a number based on given RegEx-based quantifier symbol or quantifier values.
