+++ src/modules/internet/_char-mappings.ts
