+++ src/modules/internet/module.ts
@@ -3,7 +3,7 @@ import { FakerError } from '../../errors/faker-error';
 import type { Faker } from '../../faker';
 import { toBase64Url } from '../../internal/base64';
 import { ModuleBase } from '../../internal/module-base';
-import { charMapping } from './char-mappings';
+import { charMapping } from './_char-mappings';
 
 export type EmojiType =
   | 'smiley'
