+++ src/modules/lorem/module.ts
@@ -1,6 +1,6 @@
 import { ModuleBase } from '../../internal/module-base';
 import type { NumberRange } from '../../utils/types';
-import { filterWordListByLength } from '../word/filter-word-list-by-length';
+import { filterWordListByLength } from '../word/_filter-word-list-by-length';
 
 /**
  * Module to generate random texts and words.
