+++ src/modules/string/_uuid.ts
