+++ src/modules/word/module.ts
@@ -1,7 +1,7 @@
 import { FakerError } from '../../errors/faker-error';
 import { ModuleBase } from '../../internal/module-base';
 import type { NumberRange } from '../../utils/types';
-import { filterWordListByLength } from './filter-word-list-by-length';
+import { filterWordListByLength } from './_filter-word-list-by-length';
 
 /**
  * Module to return various types of words.
