+++ test/modules/finance-iban.spec.ts
@@ -2,7 +2,7 @@ import { isIBAN } from 'validator';
 import { describe, expect, it } from 'vitest';
 import { faker } from '../../src';
 import { prettyPrintIban } from '../../src/modules/finance';
-import ibanLib from '../../src/modules/finance/iban';
+import ibanLib from '../../src/modules/finance/_iban';
 import { times } from '../support/times';
 
 const NON_SEEDED_BASED_RUN = 25;
