+++ test/modules/finance.spec.ts
@@ -7,9 +7,9 @@ import { FakerError } from '../../src/errors/faker-error';
 import {
   BitcoinAddressFamily,
   BitcoinNetwork,
-} from '../../src/modules/finance/bitcoin';
-import ibanLib from '../../src/modules/finance/iban';
-import { luhnCheck } from '../../src/modules/helpers/luhn-check';
+} from '../../src/modules/finance/_bitcoin';
+import ibanLib from '../../src/modules/finance/_iban';
+import { luhnCheck } from '../../src/modules/helpers/_luhn-check';
 import { seededTests } from '../support/seeded-runs';
 import { times } from '../support/times';
