+++ test/modules/phone.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { faker, fakerEN_GB } from '../../src';
-import { luhnCheck } from '../../src/modules/helpers/luhn-check';
+import { luhnCheck } from '../../src/modules/helpers/_luhn-check';
 import { seededTests } from '../support/seeded-runs';
 import { times } from '../support/times';
