+++ test/modules/word.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { faker } from '../../src';
-import { filterWordListByLength } from '../../src/modules/word/filter-word-list-by-length';
+import { filterWordListByLength } from '../../src/modules/word/_filter-word-list-by-length';
 import { seededTests } from '../support/seeded-runs';
 import { times } from '../support/times';
